Avoid race condition in angular bootstrap
Currently, Chrome computes the height of a popup window by
evaluating the size of the body and animating the window to be that
size. If angular is initializing during this process, Chrome will
compute an invalid size for the popup window. By delaying the
bootstrap process until after the event thread runs, we cause
Chrome to re-render and re-animate the popup, fixing the issue
where the popup shows with the top half truncated.